Cleaned up some code.
