+++ BucketLists/BucketList.swift
@@ -36,10 +36,10 @@ struct BucketList: Hashable, Codable {
         
     ], color: Color(uiColor: .red)), BucketList(owner: "Chris's List", items: [
                                    
-                                                    Item(name: "Go to Japan", description: "Take a trip to japan and eat sushi", location: "Japan", goalDate: Date(), isComplete: false, details: "", imageArray: []),
-                                                    Item(name: "Go to Germany", description: "Take a trip to Germany", location: "Germany", goalDate: Date(), isComplete: false, details: "", imageArray: []),
-                                                    Item(name: "The the Grand Canyon", description: "Plant a trip to go visit the Gran Canyon someday", location: "Grand Canyon", goalDate: Date(), isComplete: true, details: "", imageArray: []),
-                                                    Item(name: "Finish school", description: "Finish going to school", location: "Grand Canyon", goalDate: Date(), isComplete: true, details: "", imageArray: [])],
+                                                    Item(name: "Go to Japan", description: "Take a trip to japan and eat sushi", location: nil, goalDate: Date(), isComplete: false, details: "", imageArray: []),
+                                                    Item(name: "Go to Germany", description: "Take a trip to Germany", location: nil, goalDate: Date(), isComplete: false, details: "", imageArray: []),
+                                                    Item(name: "The the Grand Canyon", description: "Plant a trip to go visit the Gran Canyon someday", location: nil, goalDate: Date(), isComplete: true, details: "", imageArray: []),
+                                                    Item(name: "Finish school", description: "Finish going to school", location: nil, goalDate: Date(), isComplete: true, details: "", imageArray: [])],
                                  
        color: Color(uiColor: .blue)), BucketList(owner: "Jake's List", items: [], color: Color(uiColor: .yellow))]
 }
@@ -59,6 +59,11 @@ struct Item: Hashable, Codable {
     var imageArray: [String]
 }
 
+struct Location: Codable, Hashable {
+    var latitude: String
+    var longitude: String
+    var location: String
+}
 
 struct Color : Codable, Hashable {
     var red : CGFloat = 0.0, green: CGFloat = 0.0, blue: CGFloat = 0.0, alpha: CGFloat = 0.0
